refactor(observability): extract foundation-pure modules to proximadb-observability-engine (Slice 1, ~10K root reduction) (#878)
* refactor(observability): extract foundation-pure modules to proximadb-observability-engine (Slice 1)

Slice 1 of the observability root-crate decomposition: move the 18
foundation-pure observability modules out of the monolithic root crate into a
new crates/modalities/proximadb-observability-engine crate.

Durable OOM relief: those modules' ~10K LOC (including their inline
#[cfg(test)] suites) now compile into the engine crate's own test binary
instead of the root's ~940K-LOC test binary, whose link step intermittently
OOMs the 16GB CI runner (exit 143). The root re-exports the crate, so all
~82 inbound callers (`crate::observability::*`) are unchanged.

Scope (data-driven, cross-checked against the .victor/project.db tree-sitter
graph — graph CALLS edges are trait-impl devirtualization noise; compile-time
grep is authoritative):
- Moved: alerting/, audit/, io_trace, metering_event, metering_otlp,
  object_store_trace, precision_metrics, predicate_diagnostics, route_explain,
  search_plan_trace, search_plan_trace_builder, tenant_label, trace_batcher,
  trace_digest, trace_fingerprint, trace_retention, trace_sampling, workload_mix.
- Kept in root (storage/rank-coupled, deferred to Slice 2): mod.rs,
  graph_linking, ingestion/, query/, storage/, rank_metrics (rank-onnx coupling).
- Foundation-clean: depends only on proximadb-{data-model,kernel,records,tenant}
  + horizontal crates; no upward edge into root storage/query/services.
  Storage up-edge ratchet unchanged at 197.

Slice 2 will land the coupled core (ingestion/query/mod.rs) behind an
ObservabilityStoragePort (dependency inversion for the WAL-backed
ObservabilityStorage, whose root-storage up-edges block a whole-module move).

Verified locally: cargo check --lib + --tests (root + new crate) clean; fmt; layering.

* fix(observability-engine): declare io-trace feature + clippy allow

CI clippy (-D warnings) failed on the new crate: `unexpected cfg condition
value: io-trace` (7×) — io_trace.rs gates sections behind `#[cfg(feature =
"io-trace")]`, but the crate didn't declare that feature, so check-cfg (run on
the lib) treated it as unexpected. The "Rust Clippy" CI gate is `--lib --bins`,
which compiled the lib and hit the check-cfg compile error.

Fix:
- Declare `io-trace = []` in the crate (mirrors root's empty cfg flag) and wire
  root's `io-trace` feature to enable `proximadb-observability-engine/io-trace`.
- Add `#[allow(clippy::type_complexity)]` on one test-only capture cell (test
  code, not reached by CI's `--lib --bins`, but kept clean for --all-targets).

Verified: cargo clippy -p proximadb-observability-engine --lib --bins (CI form)
AND --all-targets, both with -D warnings — clean.
