+++ crates/modalities/proximadb-observability-engine/Cargo.toml
@@ -1,0 +1,65 @@
+[package]
+name = "proximadb-observability-engine"
+version.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+license.workspace = true
+authors.workspace = true
+repository.workspace = true
+homepage.workspace = true
+description = "ProximaDB observability runtime — per-query I/O trace substrate, metering, alerting, audit, and embedding/rank precision metrics (logs/metrics/traces SIEM engine). Extracted from the root crate; the storage-backed core (ingestion/query) remains in the root behind ObservabilityStoragePort (Slice 2)."
+
+[lib]
+name = "proximadb_observability_engine"
+path = "src/lib.rs"
+
+[features]
+# Mirrors the root `otlp-metering` feature: the durable per-tenant OTLP metering
+# push (metering_otlp) pulls in the opentelemetry transitive surface; a
+# size-sensitive `--no-default-features` build can drop it (the module compiles
+# to no-ops). Default-on in the root build (root enables this feature).
+otlp-metering = ["dep:opentelemetry", "dep:opentelemetry_sdk", "dep:opentelemetry-otlp"]
+# Mirrors the root `io-trace` feature: the per-query io_trace bus + the
+# ObjectStore decorator are gated behind this opt-in flag (co-design C0
+# instrumentation). Empty cfg flag — the root enables it when io-trace is on.
+io-trace = []
+
+[dependencies]
+anyhow.workspace = true
+async-trait.workspace = true
+chrono.workspace = true
+dashmap.workspace = true
+futures.workspace = true
+serde.workspace = true
+serde_json.workspace = true
+tokio.workspace = true
+tracing.workspace = true
+object_store.workspace = true
+# These are root-only deps (not yet in workspace.dependencies) — declared
+# explicitly with the same version specs as the root crate so cargo unifies
+# them. A future chore can promote them to workspace.dependencies.
+bytes = "1.5"
+hex = "0.4"
+prometheus = "0.14"
+uuid = { version = "1.0", features = ["v4", "serde"] }
+# Alerting notification channels (email + webhook) — root-only deps, same specs.
+lettre = "0.11"
+reqwest = { version = "0.11", features = ["json"] }
+# Foundation types — the pure observability modules are foundation-clean
+# (they name data_model/kernel/records/tenant types, not proto messages and
+# not root storage/query/services).
+proximadb-data-model.workspace = true
+proximadb-kernel.workspace = true
+proximadb-records.workspace = true
+proximadb-tenant.workspace = true
+# Optional, feature-gated (otlp-metering) — the OTLP metering *export* surface.
+# Same specs as the root crate's optional deps.
+opentelemetry = { version = "0.32", optional = true }
+opentelemetry_sdk = { version = "0.32", features = ["rt-tokio", "experimental_metrics_periodicreader_with_async_runtime", "metrics"], optional = true }
+opentelemetry-otlp = { version = "0.32", default-features = false, features = ["grpc-tonic", "metrics"], optional = true }
+
+[dev-dependencies]
+tempfile.workspace = true
+
+[lints]
+workspace = true
