+++ crates/modalities/proximadb-observability-engine/src/search_plan_trace.rs
