+++ crates/modalities/proximadb-observability-engine/src/trace_digest.rs
